fix sharing violation on startup
copy files earlier so they don't interfere with UI
